requirements/common.txt
signalslot
python-vlc
luma.core
luma.oled
Encoder
python-mpd2
pillow
pyqt5